+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## Features
 
 **Frotend:**
+
 - Added a "Poke-A-Byte settings" panel to the settings tab.
   Allows configuration of the `appsettings.json` options through the Web UI.
 - Added a "UI settings" panel to the settings tab.
@@ -15,12 +16,14 @@
 - Added favorites 
   - You can favorite a mapper via the heart icon in the mapper select.
   - Favorite mappers are shown as buttons above the "Load mapper" panel for those mappers with one click.
+  - In the UI settings panel, you can clear your favorites or change their order.
 
+## Bugfixes
 
-## General changes
+## Other changes
 
-- Put the GitHub settings in the frontend in a box similar to the boxes on the mapper page.
-- Minor frontend UI tweaks.
+- Minor UI tweaks (phrasing, margins, etc).
+
 
 # 0.9.3